Move files to animationsonly
