+++ common/src/main/java/net/bettercombat/mixin/animationsonly/LivingEntityAnimationsOnlyMixin.java
@@ -1,4 +1,4 @@
-package net.bettercombat.mixin;
+package net.bettercombat.mixin.animationsonly;
 
 import net.minecraft.entity.LivingEntity;
 import org.spongepowered.asm.mixin.Mixin;
