+++ common/src/main/java/net/bettercombat/mixin/animationsonly/PlayerEntityAnimationsOnlyMixin.java
@@ -1,6 +1,5 @@
-package net.bettercombat.mixin;
+package net.bettercombat.mixin.animationsonly;
 
-import com.mojang.logging.LogUtils;
 import net.bettercombat.BetterCombat;
 import net.bettercombat.client.animation.PlayerAttackAnimatable;
 import net.bettercombat.logic.AnimatedHand;
@@ -13,7 +12,6 @@ import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
 import net.minecraft.util.hit.HitResult;
 import net.minecraft.util.math.Box;
-import org.slf4j.Logger;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
@@ -22,8 +20,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(PlayerEntity.class)
 public abstract class PlayerEntityAnimationsOnlyMixin extends LivingEntityAnimationsOnlyMixin {
-    private final Logger LOGGER = LogUtils.getLogger();
-
     @Unique private int comboCount = 0;
 
     @Unique private int ticksToResetCombo = 0;
